[hand-control] move hand-based smach code into another file
